backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
Pillow==10.1.0
pillow-heif==0.13.1

# Utilities
structlog==23.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Performance
aiofiles==23.2.1
httpx==0.25.2

# ML inference (optional, for AI features)
onnxruntime==1.16.3
numpy==1.26.2